+++ npm/vue/examples/nuxt/components/Logo.spec.ts
@@ -1,0 +1,9 @@
+import { mount } from '@cypress/vue'
+import Logo from './Logo.vue'
+
+describe('<Logo />', () => {
+  it('contains an svg', () => {
+    mount(Logo)
+    cy.get('svg').should('be.visible')
+  })
+})