+++ npm/design-system/CHANGELOG.md
@@ -1,3 +1,30 @@
+# [@cypress/design-system-v1.2.0](https://github.com/cypress-io/cypress/compare/@cypress/design-system-v1.1.0...@cypress/design-system-v1.2.0) (2021-04-05)
+
+
+### Bug Fixes
+
+* remove failing imports ([#15661](https://github.com/cypress-io/cypress/issues/15661)) ([357a296](https://github.com/cypress-io/cypress/commit/357a296fde0355fda6388345b1d01e4ffa11d69c))
+* **runner-ct:** add highlighting to spec list fuzzy find ([#15604](https://github.com/cypress-io/cypress/issues/15604)) ([56234e5](https://github.com/cypress-io/cypress/commit/56234e52d6d1cbd292acdfd5f5d547f0c4706b51))
+
+
+### Features
+
+* simplify vite server ([#15416](https://github.com/cypress-io/cypress/issues/15416)) ([adc2fc8](https://github.com/cypress-io/cypress/commit/adc2fc893fbf32f1f6121d18ddb8a8096e5ebf39))
+* **deps:** update dependency electron to version 12.x 🌟 ([#15292](https://github.com/cypress-io/cypress/issues/15292)) ([b52ac98](https://github.com/cypress-io/cypress/commit/b52ac98a6944bc831221ccb730f89c6cc92a4573))
+
+# [@cypress/design-system-v1.1.0](https://github.com/cypress-io/cypress/compare/@cypress/design-system-v1.0.0...@cypress/design-system-v1.1.0) (2021-03-19)
+
+
+### Bug Fixes
+
+* addressing style regressions in the CT runner ([#15542](https://github.com/cypress-io/cypress/issues/15542)) ([ecccf54](https://github.com/cypress-io/cypress/commit/ecccf5444041eeeb1aa6bc911c4e06b520182ce3))
+* specs overflow horizontally and z-index issues ([#15547](https://github.com/cypress-io/cypress/issues/15547)) ([e60a029](https://github.com/cypress-io/cypress/commit/e60a02912f9d92ae8bce9d03c2167f9a393482f6))
+
+
+### Features
+
+* **design-system:** FileExplorer  ([#15513](https://github.com/cypress-io/cypress/issues/15513)) ([f2b880c](https://github.com/cypress-io/cypress/commit/f2b880c09de5c4490027689af86b6844706c8a6b))
+
 # @cypress/design-system-v1.0.0 (2021-03-15)